pyproject.toml
[project]
name = "altium-commit"
version = "0.4.1"
description = "Interface and helper tools for Altium Designer"
readme = "README.md"
requires-python = ">=3.12"
license = {file = "LICENSE"}
authors = [{ name = "Jakub Piech", email = "piechjakub99@gmail.com" }]
keywords = ["altium", "pre-commit", "interface", "tools"]
dependencies = [
    "humanize>=4.11.0",
    "psutil>=6.1.0",
    "pyyaml>=6.0.2",
]

classifiers = [
  "Development Status :: 3 - Alpha",

  "Environment :: Console",
  
  "Intended Audience :: Manufacturing",

  "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
  "Topic :: Software Development :: Version Control :: Git",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
]

[project.urls]
Repository = "https://github.com/TMALFSSISAWBOJBUMA/altiumate.git"
Issues = "https://github.com/TMALFSSISAWBOJBUMA/altiumate/issues"

[project.scripts]
altiumate = "altiumate.main:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
"altiumate.AD_scripting" = ["*.pas", "*.PrjScr"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]
